move resolution to a new rule
