+++ spark/src/main/scala/org/apache/spark/sql/delta/PreprocessTimeTravel.scala
@@ -28,7 +28,6 @@ import org.apache.spark.sql.catalyst.rules.Rule
 import org.apache.spark.sql.catalyst.util.quoteIfNeeded
 import org.apache.spark.sql.connector.catalog.{CatalogManager, CatalogPlugin, Identifier, TableCatalog}
 import org.apache.spark.sql.connector.catalog.CatalogV2Implicits.{IdentifierHelper, MultipartIdentifierHelper}
-import org.apache.spark.sql.delta.DeltaTableUtils.{resolveCatalogAndIdentifier, resolveDeltaIdentifier}
 import org.apache.spark.sql.execution.datasources.v2.{DataSourceV2Relation, V2SessionCatalog}
 import org.apache.spark.sql.internal.{SQLConf, StaticSQLConf}
 
@@ -44,16 +43,6 @@ case class PreprocessTimeTravel(sparkSession: SparkSession) extends Rule[Logical
   override def conf: SQLConf = sparkSession.sessionState.conf
 
   override def apply(plan: LogicalPlan): LogicalPlan = plan.resolveOperators {
-    // Since the Delta's TimeTravel is a leaf node, we have to resolve the UnresolvedIdentifier.
-    case _ @ RestoreTableStatement(tt @ TimeTravel(udi: UnresolvedDeltaIdentifier, _, _, _)) =>
-      val sourceRelation = resolveIdentifier(sparkSession, udi)
-      return RestoreTableStatement(
-        TimeTravel(
-          sourceRelation,
-          tt.timestamp,
-          tt.version,
-          tt.creationSource))
-
     case ct @ CloneTableStatement(
         tt @ TimeTravel(ur: UnresolvedRelation, _, _, _), _,
           _, _, _, _, _) =>
@@ -63,22 +52,6 @@ case class PreprocessTimeTravel(sparkSession: SparkSession) extends Rule[Logical
         tt.timestamp,
         tt.version,
         tt.creationSource))
-  }
-
-  private def resolveIdentifier(
-      sparkSession: SparkSession,
-      unresolvedIdentifier: UnresolvedDeltaIdentifier): DataSourceV2Relation = {
-    val relation = resolveDeltaIdentifier(sparkSession, unresolvedIdentifier)
-    if  (relation.isDefined) {
-      relation.get
-    } else {
-      val tableId = if (unresolvedIdentifier.nameParts.length == 1) {
-        TableIdentifier(unresolvedIdentifier.nameParts.head)
-      } else {
-        unresolvedIdentifier.nameParts.asTableIdentifier
-      }
-      handleTableNotFound(sparkSession.sessionState.catalog, unresolvedIdentifier, tableId)
-    }
   }
 
   /**
@@ -100,26 +73,11 @@ case class PreprocessTimeTravel(sparkSession: SparkSession) extends Rule[Logical
       } else if (DeltaTableUtils.isValidPath(tableId)) {
         DeltaTableV2(sparkSession, new Path(tableId.table))
       } else {
-        handleTableNotFound(catalog, ur, tableId)
+        ResolveDeltaIdentifier.handleTableNotFoundInTimeTravel(catalog, ur, tableId)
       }
 
       val identifier = deltaTableV2.getTableIdentifierIfExists.map(
         id => Identifier.of(id.database.toArray, id.table))
       DataSourceV2Relation.create(deltaTableV2, None, identifier)
-  }
-
-  private def handleTableNotFound(
-      catalog: SessionCatalog,
-      inputPlan: LeafNode,
-      tableIdentifier: TableIdentifier): Nothing = {
-    if (
-      (catalog.tableExists(tableIdentifier) &&
-        catalog.getTableMetadata(tableIdentifier).tableType == CatalogTableType.VIEW) ||
-        catalog.isTempView(tableIdentifier)) {
-      // If table exists and not found to be a view, throw not supported error
-      throw DeltaErrors.notADeltaTableException("RESTORE")
-    } else {
-      inputPlan.tableNotFound(tableIdentifier.nameParts)
-    }
   }
 }
